+++ NavigationAdvancedSample/app/src/main/res/menu/bottom_nav.xml
@@ -22,7 +22,7 @@
         android:contentDescription="@string/cd_home"
         android:title="@string/title_home" />
     <item
-        android:id="@+id/list"
+        android:id="@+id/person"
         android:icon="@drawable/ic_list"
         android:contentDescription="@string/cd_list"
         android:title="@string/title_list" />
