--- NavigationAdvancedSample/app/src/main/res/navigation/list.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright 2019, The Android Open Source Project
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~      http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<navigation
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/list"
-    app:startDestination="@+id/leaderboard">
-
-    <fragment
-            android:id="@+id/leaderboard"
-            android:name="com.example.android.navigationadvancedsample.listscreen.Leaderboard"
-            android:label="@string/title_list"
-            tools:layout="@layout/fragment_leaderboard">
-        <action
-            android:id="@+id/action_leaderboard_to_userProfile"
-            app:destination="@id/userProfile"/>
-    </fragment>
-    <fragment
-        android:id="@+id/userProfile"
-        android:name="com.example.android.navigationadvancedsample.dynamicfeature1.listscreen.UserProfile"
-        android:label="@string/title_detail"
-        tools:layout="@layout/fragment_user_profile">
-        <deepLink
-            android:id="@+id/deepLink"
-            app:uri="www.example.com/user/{userName}"
-            android:autoVerify="true"/>
-        <argument
-            android:name="userName"
-            app:argType="string"/>
-    </fragment>
-</navigation>
