+++ NavigationAdvancedSample/app/src/main/res/values/strings.xml
@@ -48,4 +48,5 @@
     <string name="cd_list">List</string>
     <string name="cd_form">Form</string>
     <string name="title_dynamicfeature1">dynamic-feature-1</string>
+    <string name="title_personfeature">PersonFeature</string>
 </resources>
