+++ NavigationAdvancedSample/build.gradle
@@ -20,6 +20,7 @@ buildscript {
     apply from: 'versions.gradle'
     ext {
         versions.kotlin = '1.4.20'
+        versions.kotlin = '1.4.20'
     }
     addRepos(repositories)
     dependencies {
