--- NavigationAdvancedSample/dynamicfeature1/build.gradle
@@ -1,47 +1,0 @@
-plugins {
-    id 'com.android.dynamic-feature'
-    id 'kotlin-android'
-}
-android {
-    compileSdkVersion 30
-    buildToolsVersion "30.0.2"
-
-    defaultConfig {
-        applicationId "com.example.android.navigationadvancedsample.dynamicfeature1"
-        minSdkVersion 23
-        targetSdkVersion 30
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    implementation project(":app")
-    implementation deps.kotlin.stdlib
-    implementation deps.app_compat
-    implementation deps.material
-    implementation deps.core_ktx
-    implementation deps.constraint_layout
-    implementation deps.arch_core.runtime
-
-    // Navigation
-    implementation deps.navigation.runtime_ktx
-    implementation deps.navigation.fragment_ktx
-    implementation deps.navigation.ui_ktx
-    implementation deps.navigation.dfm
-
-
-    testImplementation 'junit:junit:4.+'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
-    androidTestImplementation 'androidx.annotation:annotation:1.1.0'
-}
