--- NavigationAdvancedSample/dynamicfeature1/src/main/java/com/example/android/navigationadvancedsample/dynamicfeature1/FeatureNextFragment.kt
@@ -1,17 +1,0 @@
-package com.example.android.navigationadvancedsample.dynamicfeature1
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-
-
-class FeatureNextFragment : Fragment() {
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_feature_next, container, false)
-    }
-}
