+++ NavigationAdvancedSample/features/PersonFeature/src/main/res/layout/fragment_leaderboard.xml
