+++ NavigationAdvancedSample/features/PersonFeature/src/main/res/layout/fragment_user_profile.xml
