+++ NavigationAdvancedSample/features/PersonFeature/src/main/res/layout/list_view_item.xml
