+++ NavigationAdvancedSample/features/PersonFeature/src/test/java/com/example/android/navigationadvancedsample/PersonFeature/ExampleUnitTest.kt
@@ -1,0 +1,17 @@
+package com.example.android.navigationadvancedsample.PersonFeature
+
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
