+++ NavigationAdvancedSample/features/dynamicfeature1/.gitignore
@@ -1,0 +1,1 @@
+/build
