+++ NavigationAdvancedSample/features/dynamicfeature1/src/main/AndroidManifest.xml
