+++ NavigationAdvancedSample/features/dynamicfeature1/src/main/java/com/example/android/navigationadvancedsample/dynamicfeature1/FeatureFragment.kt
