+++ NavigationAdvancedSample/features/dynamicfeature1/src/main/res/layout/fragment_feature.xml
