+++ NavigationAdvancedSample/features/dynamicfeature1/src/main/res/layout/fragment_feature_next.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    android:orientation="vertical"
+    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="next is here"
+        />
+
+    <Button
+        android:id="@+id/person_list_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="To person list"/>
+</LinearLayout>
