+++ NavigationAdvancedSample/features/dynamicfeature1/src/main/res/navigation/dynamic_list_graph.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright 2019, The Android Open Source Project
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<navigation
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@id/dynamic_list_graph"
+    app:moduleName="dynamicfeature1"
+    app:startDestination="@id/feature">
+
+    <fragment
+            android:id="@+id/feature"
+            android:name="com.example.android.navigationadvancedsample.dynamicfeature1.FeatureFragment"
+            android:label="feature"
+            tools:layout="@layout/fragment_feature">
+        <action
+            android:id="@+id/action_feature_to_feature_next"
+            app:destination="@id/feature_next"/>
+    </fragment>
+    <fragment
+        android:id="@+id/feature_next"
+        android:name="com.example.android.navigationadvancedsample.dynamicfeature1.FeatureNextFragment"
+        android:label="FeatureNext"
+        tools:layout="@layout/fragment_feature_next">
+        <action
+            android:id="@+id/action_feature_next_to_person_list"
+            app:destination="@id/list"/>
+
+    </fragment>
+
+    <include-dynamic
+        android:id="@id/list"
+        app:graphPackage="com.example.android.navigationadvancedsample.PersonFeature"
+        app:graphResName="list"
+        app:moduleName="PersonFeature"/>
+
+</navigation>
