+++ NavigationAdvancedSample/settings.gradle
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-include ':dynamicfeature1'
+include ':features:PersonFeature'
+include ':features:dynamicfeature1'
 
 include ':app'
